Rename data block to repeater block
