+++ packages/client/src/components/app/blocks/index.js
@@ -1,3 +1,3 @@
 export { default as tableblock } from "./TableBlock.svelte"
 export { default as cardsblock } from "./CardsBlock.svelte"
-export { default as datablock } from "./DataBlock.svelte"
+export { default as repeaterblock } from "./RepeaterBlock.svelte"
